feat: remove CardDot
