+++ src/shared/components/EmojiDisplay/EmojiDisplay.tsx
@@ -1,0 +1,7 @@
+import { EmojiDisplayContainer } from "shared/components/EmojiDisplay/emoji-display.style";
+
+const EmojiDisplay = ({ src }: { src: string }) => (
+	<EmojiDisplayContainer src={src} />
+);
+
+export default EmojiDisplay;
